Connected graph using BFS algorithm
